Hint Function works!!
The "Hint" function now works, and I've added some more comments to make it easier to understand the code later.
The code is messy now, and I really want to fix it so it's more clear to understand, but i'll leave that work out for now, since this is just for practice. Maybe I'll comeback much later.
I might move on to other projects.